Move app hash logic from kube interface to add command
